feat: implement lazy loading for CLI commands to improve startup performance
- Optimize checkNodeVersion by removing heavy imports (~150ms savings)
- Refactor command files to use minimal imports pattern
- Move heavy logic to separate action files that are dynamically imported
- Commands refactored: eval, init, view, generate, share, show, list, cache
- Add benchmark scripts to measure CLI startup performance
- Initial results: ~68ms improvement (977ms -> 909ms)

Further optimizations possible by lazy loading command registration in main.ts
