+++ src/commands/generate/assertions.ts
@@ -1,122 +1,16 @@
-import chalk from 'chalk';
 import type { Command } from 'commander';
 import { InvalidArgumentError } from 'commander';
-import * as fs from 'fs';
-import yaml from 'js-yaml';
-import { synthesizeFromTestSuite } from '../../assertions/synthesis';
-import { disableCache } from '../../cache';
-import logger from '../../logger';
-import telemetry from '../../telemetry';
-import { type TestSuite, type UnifiedConfig } from '../../types';
-import { isRunningUnderNpx, printBorder, setupEnv } from '../../util';
-import { resolveConfigs } from '../../util/config/load';
+import { type UnifiedConfig } from '../../types';
 
-interface DatasetGenerateOptions {
-  cache: boolean;
-  config?: string;
-  envFile?: string;
-  instructions?: string;
-  numAssertions?: string;
-  output?: string;
-  provider?: string;
-  write: boolean;
-  defaultConfig: Partial<UnifiedConfig>;
-  defaultConfigPath: string | undefined;
-  type: 'pi' | 'g-eval' | 'llm-rubric';
-}
+// Re-export for backward compatibility
+export { doGenerateAssertions } from './actions/assertionsAction';
 
-export async function doGenerateAssertions(options: DatasetGenerateOptions): Promise<void> {
-  setupEnv(options.envFile);
-  if (!options.cache) {
-    logger.info('Cache is disabled.');
-    disableCache();
-  }
-
-  let testSuite: TestSuite;
-  const configPath = options.config || options.defaultConfigPath;
-  if (configPath) {
-    const resolved = await resolveConfigs(
-      {
-        config: [configPath],
-      },
-      options.defaultConfig,
-      'AssertionGeneration',
+function validateAssertionType(value: string) {
+  const validTypes = ['pi', 'g-eval', 'llm-rubric'];
+  if (!validTypes.includes(value)) {
+    throw new InvalidArgumentError(
+      `Invalid assertion type. Must be one of: ${validTypes.join(', ')}`,
     );
-    testSuite = resolved.testSuite;
-  } else {
-    throw new Error('Could not find config file. Please use `--config`');
-  }
-
-  const startTime = Date.now();
-  telemetry.record('command_used', {
-    name: 'generate_assertions - started',
-    numPrompts: testSuite.prompts.length,
-    numTestsExisting: (testSuite.tests || []).length,
-  });
-
-  const results = await synthesizeFromTestSuite(testSuite, {
-    instructions: options.instructions,
-    numQuestions: Number.parseInt(options.numAssertions || '5', 10),
-    provider: options.provider,
-    type: options.type,
-  });
-  const configAddition = {
-    assert: results,
-  };
-  const yamlString = yaml.dump(configAddition);
-  if (options.output) {
-    // Should the output be written as a YAML or CSV?
-    if (options.output.endsWith('.yaml')) {
-      fs.writeFileSync(options.output, yamlString);
-    } else {
-      throw new Error(`Unsupported output file type: ${options.output}`);
-    }
-    printBorder();
-    logger.info(`Wrote ${results.length} new assertions to ${options.output}`);
-    printBorder();
-  } else {
-    printBorder();
-    logger.info('New test Cases');
-    printBorder();
-    logger.info(yamlString);
-  }
-
-  printBorder();
-  if (options.write && configPath) {
-    const existingConfig = yaml.load(fs.readFileSync(configPath, 'utf8')) as Partial<UnifiedConfig>;
-    // Handle the union type for tests (string | TestGeneratorConfig | Array<...>)
-    const existingDefaultTest =
-      typeof existingConfig.defaultTest === 'object' ? existingConfig.defaultTest : {};
-    existingConfig.defaultTest = {
-      ...existingDefaultTest,
-      assert: [...(existingDefaultTest?.assert || []), ...configAddition.assert],
-    };
-    fs.writeFileSync(configPath, yaml.dump(existingConfig));
-    logger.info(`Wrote ${results.length} new test cases to ${configPath}`);
-    const runCommand = isRunningUnderNpx() ? 'npx promptfoo eval' : 'promptfoo eval';
-    logger.info(chalk.green(`Run ${chalk.bold(runCommand)} to run the generated assertions`));
-  } else {
-    logger.info(
-      `Copy the above test cases or run ${chalk.greenBright(
-        'promptfoo generate assertions --write',
-      )} to write directly to the config`,
-    );
-  }
-
-  telemetry.record('command_used', {
-    duration: Math.round((Date.now() - startTime) / 1000),
-    name: 'generate_assertions',
-    numPrompts: testSuite.prompts.length,
-    numTestsExisting: (testSuite.tests || []).length,
-    numAssertionsGenerated: results.length,
-    provider: options.provider || 'default',
-  });
-}
-
-function validateAssertionType(value: string, previous: string) {
-  const allowedStrings = ['pi', 'g-eval', 'llm-rubric'];
-  if (!allowedStrings.includes(value)) {
-    throw new InvalidArgumentError(`Option --type must be one of: ${allowedStrings.join(', ')}.`);
   }
   return value;
 }
@@ -144,12 +38,13 @@ export function generateAssertionsCommand(
     .option('--numAssertions <amount>', 'Number of assertions to generate')
     .option(
       '--provider <provider>',
-      `Provider to use for generating assertions. Defaults to the default grading provider.`,
-    )
-    .option(
-      '-i, --instructions [instructions]',
-      'Additional instructions to follow while generating assertions',
+      `Provider to use for generating adversarial tests. Defaults to the default grading provider.`,
     )
     .option('--no-cache', 'Do not read or write results to disk cache', false)
-    .action((opts) => doGenerateAssertions({ ...opts, defaultConfig, defaultConfigPath }));
+    .option('--env-file, --env-path <path>', 'Path to .env file')
+    .action(async (opts) => {
+      // Lazy load the action handler
+      const { doGenerateAssertions } = await import('./actions/assertionsAction');
+      await doGenerateAssertions({ ...opts, defaultConfig, defaultConfigPath });
+    });
 }
