+++ src/commands/view.ts
@@ -1,11 +1,5 @@
 import type { Command } from 'commander';
 import { getDefaultPort } from '../constants';
-import logger from '../logger';
-import { startServer } from '../server/server';
-import telemetry from '../telemetry';
-import { setupEnv } from '../util';
-import { setConfigDirectoryPath } from '../util/config/manage';
-import { BrowserBehavior } from '../util/server';
 
 export function viewCommand(program: Command) {
   program
@@ -16,39 +10,9 @@ export function viewCommand(program: Command) {
     .option('-n, --no', 'Skip confirmation and do not open the URL')
     .option('--filter-description <pattern>', 'Filter evals by description using a regex pattern')
     .option('--env-file, --env-path <path>', 'Path to .env file')
-    .action(
-      async (
-        directory: string | undefined,
-        cmdObj: {
-          port: number;
-          yes: boolean;
-          no: boolean;
-          apiBaseUrl?: string;
-          envPath?: string;
-          filterDescription?: string;
-        } & Command,
-      ) => {
-        setupEnv(cmdObj.envPath);
-        telemetry.record('command_used', {
-          name: 'view',
-        });
-
-        if (cmdObj.filterDescription) {
-          logger.warn(
-            'The --filter-description option is deprecated and not longer supported. The argument will be ignored.',
-          );
-        }
-
-        if (directory) {
-          setConfigDirectoryPath(directory);
-        }
-        // Block indefinitely on server
-        const browserBehavior = cmdObj.yes
-          ? BrowserBehavior.OPEN
-          : cmdObj.no
-            ? BrowserBehavior.SKIP
-            : BrowserBehavior.ASK;
-        await startServer(cmdObj.port, browserBehavior);
-      },
-    );
+    .action(async (directory: string | undefined, cmdObj: any) => {
+      // Lazy load the action handler
+      const { viewAction } = await import('./view/viewAction');
+      await viewAction(directory, cmdObj);
+    });
 }
